# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: KOE Backend API
  description: 'Event Sourcing API for NS 8407 change order claims (Krav om Endringsordre).


    ## Authentication

    - **CSRF Token**: Required for POST/PUT/DELETE via `X-CSRF-Token` header

    - **Magic Link**: Required for case operations via `Authorization: Bearer <token>`
    header


    ## Event Sourcing

    All state changes are recorded as immutable events. State is computed by replaying
    events.

    Optimistic concurrency control via `expected_version` prevents conflicts.'
  version: 1.0.0
  contact:
    name: KOE Development Team
servers:
- url: http://localhost:8080
  description: Local development
...
paths:
...
  /api/forsering/opprett:
    post:
      tags:
      - Forsering
      summary: Create forsering case
      description: 'Create a new forsering (acceleration) case based on rejected deadline
        extensions.


        Per NS 8407 §33.8: When BH rejects a justified deadline extension claim,

        TE can treat it as an order to accelerate if the cost is within 30% of

        the liquidated damages that would have accrued.'
      operationId: createForsering
      security:
      - csrfToken: []
      - magicLink: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - avslatte_sak_ids
              - estimert_kostnad
              - dagmulktsats
              - begrunnelse
              properties:
                avslatte_sak_ids:
                  type: array
                  items:
                    type: string
                  description: IDs of rejected deadline extension cases
                estimert_kostnad:
                  type: number
                  description: Estimated acceleration cost
                dagmulktsats:
                  type: number
                  description: Liquidated damages rate per day
                begrunnelse:
                  type: string
                  description: Justification
                avslatte_dager:
                  type: integer
                  description: Total rejected days (auto-calculated if omitted)
      responses:
        '201':
          description: Forsering case created
        '400':
          description: Validation error (e.g., cost exceeds 30% limit)
...
  /api/forsering/{sak_id}/kontekst:
    get:
      tags:
      - Forsering
      summary: Get forsering context
      description: Get complete context including related cases, states, and events
      operationId: getForseringKontekst
      security:
      - magicLink: []
      parameters:
      - name: sak_id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Forsering context with related cases
  /api/forsering/valider:
    post:
      tags:
      - Forsering
      summary: Validate 30% rule
      description: Check if estimated cost is within the 30% limit (dagmulkt + 30%)
      operationId: validateForsering
      security:
      - magicLink: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - estimert_kostnad
              - avslatte_dager
              - dagmulktsats
              properties:
                estimert_kostnad:
                  type: number
                avslatte_dager:
                  type: integer
                dagmulktsats:
                  type: number
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  er_gyldig:
                    type: boolean
                  maks_kostnad:
                    type: number
                  prosent_av_maks:
                    type: number
  /api/forsering/{sak_id}/relaterte:
    get:
      tags:
      - Forsering
      summary: Get related cases for forsering
      description: Get all KOE cases related to a forsering case (rejected frist claims)
      operationId: getForseringRelaterte
      security:
      - magicLink: []
      parameters:
      - name: sak_id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: List of related cases
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  sak_id:
                    type: string
                  relaterte_saker:
                    type: array
                    items:
                      type: object
                      properties:
                        relatert_sak_id:
                          type: string
                        relatert_sak_tittel:
                          type: string
                        bimsync_issue_board_ref:
                          type: string
                        bimsync_issue_number:
                          type: integer
  /api/forsering/{sak_id}/relatert:
    post:
      tags:
      - Forsering
      summary: Add related case to forsering
      description: Add a KOE case as related to the forsering
      operationId: addForseringRelatert
      security:
      - csrfToken: []
      - magicLink: []
      parameters:
      - name: sak_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - koe_sak_id
              properties:
                koe_sak_id:
                  type: string
                  description: ID of KOE case to add
      responses:
        '200':
          description: KOE added to forsering
  /api/forsering/{sak_id}/relatert/{koe_sak_id}:
    delete:
      tags:
      - Forsering
      summary: Remove related case from forsering
      operationId: removeForseringRelatert
      security:
      - csrfToken: []
      - magicLink: []
      parameters:
      - name: sak_id
        in: path
        required: true
        schema:
          type: string
      - name: koe_sak_id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: KOE removed from forsering
  /api/forsering/{sak_id}/bh-respons:
    post:
      tags:
      - Forsering
      summary: Register BH response to forsering
      description: BH accepts or rejects the forsering claim
      operationId: bhResponsForsering
      security:
      - csrfToken: []
      - magicLink: []
      parameters:
      - name: sak_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - aksepterer
              - begrunnelse
              properties:
                aksepterer:
                  type: boolean
                  description: Whether BH accepts the forsering
                godkjent_kostnad:
                  type: number
                  description: Approved cost (may be lower than estimated)
                begrunnelse:
                  type: string
                  description: Justification for decision
                expected_version:
                  type: integer
                  description: Expected version for optimistic concurrency
      responses:
        '200':
          description: BH response registered
          content:
            application/json:
              schema:
                allOf:
                - type: object
                  properties:
                    success:
                      type: boolean
                    message:
                      type: string
                    state:
                      type: object
                    version:
                      type: integer
                - $ref: '#/components/schemas/CatendaSyncStatus'
        '409':
          description: Concurrency conflict - version mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConcurrencyConflictError'
  /api/forsering/{sak_id}/stopp:
    post:
      tags:
      - Forsering
      summary: Stop active forsering
      description: Stop an ongoing forsering and record incurred costs
      operationId: stoppForsering
      security:
      - csrfToken: []
      - magicLink: []
      parameters:
      - name: sak_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - begrunnelse
              properties:
                begrunnelse:
                  type: string
                  description: Reason for stopping
                paalopte_kostnader:
                  type: number
                  description: Incurred costs at time of stop
                expected_version:
                  type: integer
                  description: Expected version for optimistic concurrency
      responses:
        '200':
          description: Forsering stopped
          content:
            application/json:
              schema:
                allOf:
                - type: object
                  properties:
                    success:
                      type: boolean
                    message:
                      type: string
                    dato_stoppet:
                      type: string
                      format: date
                    state:
                      type: object
                    version:
                      type: integer
                - $ref: '#/components/schemas/CatendaSyncStatus'
        '409':
          description: Concurrency conflict - version mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConcurrencyConflictError'
  /api/forsering/{sak_id}/kostnader:
    put:
      tags:
      - Forsering
      summary: Update incurred costs
      description: Update the incurred costs for an active forsering
      operationId: oppdaterForseringKostnader
      security:
      - csrfToken: []
      - magicLink: []
      parameters:
      - name: sak_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - paalopte_kostnader
              properties:
                paalopte_kostnader:
                  type: number
                  description: Current incurred costs
                kommentar:
                  type: string
                  description: Optional comment on update
                expected_version:
                  type: integer
                  description: Expected version for optimistic concurrency
      responses:
        '200':
          description: Costs updated
          content:
            application/json:
              schema:
                allOf:
                - type: object
                  properties:
                    success:
                      type: boolean
                    message:
                      type: string
                    state:
                      type: object
                    version:
                      type: integer
                - $ref: '#/components/schemas/CatendaSyncStatus'
        '409':
          description: Concurrency conflict - version mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConcurrencyConflictError'
  /api/forsering/by-relatert/{sak_id}:
    get:
      tags:
      - Forsering
      summary: Find forsering cases referencing a KOE
      description: Find all forsering cases that reference a given KOE case. Used
        for back-links.
      operationId: findForseringerForSak
      security:
      - magicLink: []
      parameters:
      - name: sak_id
        in: path
        required: true
        schema:
          type: string
        description: KOE case ID to search for
      responses:
        '200':
          description: List of forsering cases referencing this KOE
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  forseringer:
                    type: array
                    items:
                      type: object
                      properties:
                        forsering_sak_id:
                          type: string
                        tittel:
                          type: string
                        status:
                          type: string
                        estimert_kostnad:
                          type: number
...
components:
  securitySchemes:
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
    magicLink:
      type: http
      scheme: bearer
  schemas:
...
    ConcurrencyConflictError:
      type: object
      description: Returned when optimistic concurrency check fails (409)
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: CONCURRENCY_CONFLICT
        message:
          type: string
          example: 'Versjonskonflikt: forventet 3, fikk 4'
        expected_version:
          type: integer
        actual_version:
          type: integer
    CatendaSyncStatus:
      type: object
      description: Status for Catenda synchronization (comment + status update)
      properties:
        catenda_synced:
          type: boolean
          description: Whether the event was successfully synced to Catenda
        catenda_comment_posted:
          type: boolean
          description: Whether a comment was posted to the Catenda topic
        catenda_status_updated:
          type: boolean
          description: Whether the topic status was updated in Catenda
        catenda_skipped_reason:
          type: string
          enum:
          - no_topic_id
          - not_authenticated
          - no_client
          - sync_not_attempted
          - error
          nullable: true
          description: Reason why sync was skipped (if catenda_synced=false)
        catenda_error:
          type: string
          nullable: true
          description: Error message if sync failed
...